+++ Source/Blazorise.Bootstrap/BootstrapClassProvider.cs
@@ -453,20 +453,6 @@ namespace Blazorise.Bootstrap
 
         #endregion
 
-        #region Snackbar
-
-        public virtual string Snackbar() => "snackbar";
-
-        public virtual string SnackbarShow() => Show();
-
-        public virtual string SnackbarMultiline() => "snackbar-multi-line";
-
-        public virtual string SnackbarBody() => "snackbar-body";
-
-        public virtual string SnackbarAction() => "snackbar-btn";
-
-        #endregion
-
         #region Alert
 
         public virtual string Alert() => "alert";