+++ Source/Blazorise.Bulma/BulmaClassProvider.cs
@@ -467,20 +467,6 @@ namespace Blazorise.Bulma
 
         #endregion
 
-        #region Snackbar
-
-        public virtual string Snackbar() => "snackbar";
-
-        public virtual string SnackbarShow() => Show();
-
-        public virtual string SnackbarMultiline() => "snackbar-multi-line";
-
-        public virtual string SnackbarBody() => "snackbar-body";
-
-        public virtual string SnackbarAction() => "snackbar-btn";
-
-        #endregion
-
         #region Alert
 
         public virtual string Alert() => "notification";